requirements.txt
flask>=2.0.0
openai>=1.0.0
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
PyMuPDF
pillow
pytesseract
plotly>=5.0.0
pandas>=1.5.0
python-dateutil
firebase-admin         # Optional: for Firestore integration